% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NURS.R
\name{NURS}
\alias{NURS}
\title{NURS draws}
\source{
\url{https://arxiv.org/abs/2501.18548v2}
}
\usage{
NURS(logpdf, theta_init, n, epsilon, h, M)
}
\arguments{
\item{logpdf}{log (non-normalized) target density}

\item{theta_init}{initial state}

\item{n}{number of draws}

\item{epsilon}{non-negative density threshold}

\item{h}{positive lattice size}

\item{M}{maximum number of doublings}
}
\value{
a sequence of draws
}
\description{
NURS draws
}
