% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NURS.R
\name{NURS_step}
\alias{NURS_step}
\title{Single NURS step}
\usage{
NURS_step(logpdf, theta, epsilon, h, M)
}
\arguments{
\item{logpdf}{log (non-normalized) target density}

\item{theta}{current state}

\item{epsilon}{density threshold}

\item{h}{lattice size}

\item{M}{maximum number of doublings}
}
\value{
One NURS step from theta.
}
\description{
Single NURS step
}
