% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NURS.R
\name{NURS_stop}
\alias{NURS_stop}
\title{Check no underrun stopping condition}
\usage{
NURS_stop(log_vals, log_eps_h)
}
\arguments{
\item{log_vals}{orbit}

\item{log_eps_h}{log(epsilon) + log(h)}
}
\value{
True if Orbit satisfies stopping criterion.
}
\description{
Check no underrun stopping condition
}
