% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NURS.R
\name{NURS_sub_stop}
\alias{NURS_sub_stop}
\title{Recursively check sub stopping}
\usage{
NURS_sub_stop(log_vals, log_eps_h)
}
\arguments{
\item{log_vals}{sub orbit}

\item{log_eps_h}{log(epsilon) + log(h)}
}
\value{
True if sub orbit satisfies stopping criterion.
}
\description{
Recursively check sub stopping
}
