% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.NURS-package.R
\docType{package}
\name{R.NURS-package}
\alias{R.NURS}
\alias{R.NURS-package}
\title{R.NURS: R Implementation of No-Underrun Sampler}
\description{
A bare-bones R implementation of the No-Underrun Sampler (Nawaf Bou-Rabee, Bob Carpenter, Sifan Liu, Stefan Oberdörster. 2025.) written in base R. This implementation attempts to stick very closely to the algorithm described in Section 2 and Appendix A of the paper.
}
\author{
\strong{Maintainer}: Visruth Srimath Kandali \email{public@visruth.com}

}
\keyword{internal}
